Extract FTP tests from test_open-uri.rb
